+++ src/Audit.EntityFramework/IAuditDbContext.cs
@@ -93,10 +93,19 @@ namespace Audit.EntityFramework
         /// </summary>
         bool ReloadDatabaseValues { get; set; }
 
+#if EF_CORE
+        /// <summary>
+        /// Value to indicate whether to reload database values after saving changes.
+        /// </summary>
+        /// <remarks>Set this property to <see langword="true"/> to ensure that the entity's values are
+        /// refreshed from the database after saving changes, which may be necessary if the database generates default values or triggers.</remarks>
+        bool ReloadDatabaseValuesAfterSave { get; set; }
+#endif
+
         /// <summary>
         /// A collection of property names to include on the audit event for each entity type when using Opt-In mode.
         /// </summary>
-        Dictionary<Type, HashSet<string>> IncludedPropertyNames { get; set; }
+        Dictionary <Type, HashSet<string>> IncludedPropertyNames { get; set; }
 
         /// <summary>
         /// Value to indicate if the ChangesByColumn dictionary should be used instead of the Changes list to store the modified column values.